requirements.txt
Flask==2.2.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.2
Flask-JWT-Extended==4.4.0
Flask-Mail==0.9.1
openai==0.27.0
gunicorn==20.1.0
psycopg[binary]>=3.2.6



